requirements.txt
python-telegram-bot==20.7
speechrecognition
pydub
postgrest-py
python-dotenv
aiohttp
